requirements.txt
numpy==2.2.5
opencv-python==4.11.0.86
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
python-dotenv==1.1.0
shiboken6==6.9.0
